Refactor services related to manuf order plan
RM74672
