+++ axelor-maintenance/src/main/java/com/axelor/apps/maintenance/module/MaintenanceModule.java
@@ -28,11 +28,13 @@ import com.axelor.apps.maintenance.service.BillOfMaterialMaintenanceService;
 import com.axelor.apps.maintenance.service.BillOfMaterialServiceMaintenanceImpl;
 import com.axelor.apps.maintenance.service.MaintenanceRequestService;
 import com.axelor.apps.maintenance.service.MaintenanceRequestServiceImpl;
+import com.axelor.apps.maintenance.service.ManufOrderPlanServiceMaintenanceImpl;
 import com.axelor.apps.maintenance.service.ManufOrderPrintService;
 import com.axelor.apps.maintenance.service.ManufOrderPrintServiceImpl;
 import com.axelor.apps.maintenance.service.ManufOrderWorkflowMaintenanceServiceImpl;
 import com.axelor.apps.production.service.BillOfMaterialComputeNameServiceImpl;
 import com.axelor.apps.production.service.BillOfMaterialServiceImpl;
+import com.axelor.apps.production.service.manuforder.ManufOrderPlanServiceImpl;
 import com.axelor.apps.production.service.manuforder.ManufOrderWorkflowServiceImpl;
 
 public class MaintenanceModule extends AxelorModule {
@@ -49,5 +51,6 @@ public class MaintenanceModule extends AxelorModule {
     bind(ManufOrderPrintService.class).to(ManufOrderPrintServiceImpl.class);
     bind(BillOfMaterialComputeNameServiceImpl.class)
         .to(BillOfMaterialComputeNameServiceMaintenanceImpl.class);
+    bind(ManufOrderPlanServiceImpl.class).to(ManufOrderPlanServiceMaintenanceImpl.class);
   }
 }
