+++ axelor-maintenance/src/main/java/com/axelor/apps/maintenance/service/ManufOrderWorkflowMaintenanceServiceImpl.java
@@ -29,7 +29,6 @@ import com.axelor.apps.production.db.repo.OperationOrderRepository;
 import com.axelor.apps.production.db.repo.ProductionConfigRepository;
 import com.axelor.apps.production.service.app.AppProductionService;
 import com.axelor.apps.production.service.config.ProductionConfigService;
-import com.axelor.apps.production.service.config.StockConfigProductionService;
 import com.axelor.apps.production.service.manuforder.ManufOrderOutgoingStockMoveService;
 import com.axelor.apps.production.service.manuforder.ManufOrderOutsourceService;
 import com.axelor.apps.production.service.manuforder.ManufOrderService;
@@ -39,26 +38,21 @@ import com.axelor.apps.production.service.operationorder.OperationOrderOutsource
 import com.axelor.apps.production.service.operationorder.OperationOrderPlanningService;
 import com.axelor.apps.production.service.operationorder.OperationOrderService;
 import com.axelor.apps.production.service.operationorder.OperationOrderWorkflowService;
-import com.axelor.apps.production.service.productionorder.ProductionOrderService;
-import com.axelor.apps.purchase.service.PurchaseOrderService;
 import com.axelor.inject.Beans;
 import com.google.inject.Inject;
 import com.google.inject.persist.Transactional;
 import java.math.BigDecimal;
 import java.time.temporal.ChronoUnit;
-import org.apache.commons.collections.CollectionUtils;
 
 public class ManufOrderWorkflowMaintenanceServiceImpl extends ManufOrderWorkflowServiceImpl {
 
   @Inject
   public ManufOrderWorkflowMaintenanceServiceImpl(
       OperationOrderWorkflowService operationOrderWorkflowService,
-      OperationOrderRepository operationOrderRepo,
       ManufOrderStockMoveService manufOrderStockMoveService,
       ManufOrderRepository manufOrderRepo,
       ProductCompanyService productCompanyService,
       ProductionConfigRepository productionConfigRepo,
-      PurchaseOrderService purchaseOrderService,
       AppBaseService appBaseService,
       OperationOrderService operationOrderService,
       AppProductionService appProductionService,
@@ -67,18 +61,14 @@ public class ManufOrderWorkflowMaintenanceServiceImpl extends ManufOrderWorkflow
       ManufOrderOutgoingStockMoveService manufOrderOutgoingStockMoveService,
       ManufOrderService manufOrderService,
       SequenceService sequenceService,
-      ProductionOrderService productionOrderService,
       ManufOrderOutsourceService manufOrderOutsourceService,
-      OperationOrderOutsourceService operationOrderOutsourceService,
-      StockConfigProductionService stockConfigProductionService) {
+      OperationOrderOutsourceService operationOrderOutsourceService) {
     super(
         operationOrderWorkflowService,
-        operationOrderRepo,
         manufOrderStockMoveService,
         manufOrderRepo,
         productCompanyService,
         productionConfigRepo,
-        purchaseOrderService,
         appBaseService,
         operationOrderService,
         appProductionService,
@@ -87,56 +77,8 @@ public class ManufOrderWorkflowMaintenanceServiceImpl extends ManufOrderWorkflow
         manufOrderOutgoingStockMoveService,
         manufOrderService,
         sequenceService,
-        productionOrderService,
         manufOrderOutsourceService,
-        operationOrderOutsourceService,
-        stockConfigProductionService);
-  }
-
-  @Transactional(rollbackOn = {Exception.class})
-  @Override
-  public ManufOrder plan(ManufOrder manufOrder) throws AxelorException {
-    if (manufOrder.getTypeSelect() != ManufOrderRepository.TYPE_MAINTENANCE) {
-      return super.plan(manufOrder);
-    }
-
-    ManufOrderService manufOrderService = Beans.get(ManufOrderService.class);
-
-    if (Beans.get(SequenceService.class)
-        .isEmptyOrDraftSequenceNumber(manufOrder.getManufOrderSeq())) {
-      manufOrder.setManufOrderSeq(manufOrderService.getManufOrderSeq(manufOrder));
-    }
-
-    if (CollectionUtils.isEmpty(manufOrder.getOperationOrderList())) {
-      manufOrderService.preFillOperations(manufOrder);
-    }
-    if (!manufOrder.getIsConsProOnOperation()
-        && CollectionUtils.isEmpty(manufOrder.getToConsumeProdProductList())) {
-      manufOrderService.createToConsumeProdProductList(manufOrder);
-    }
-
-    if (CollectionUtils.isEmpty(manufOrder.getToProduceProdProductList())) {
-      manufOrderService.createToProduceProdProductList(manufOrder);
-    }
-
-    if (manufOrder.getPlannedStartDateT() == null) {
-      manufOrder.setPlannedStartDateT(
-          Beans.get(AppProductionService.class).getTodayDateTime().toLocalDateTime());
-    }
-
-    operationOrderPlanningService.plan(manufOrder.getOperationOrderList());
-
-    manufOrder.setPlannedEndDateT(this.computePlannedEndDateT(manufOrder));
-
-    if (manufOrder.getBillOfMaterial() != null) {
-      manufOrder.setUnit(manufOrder.getBillOfMaterial().getUnit());
-    }
-
-    manufOrder.setStatusSelect(ManufOrderRepository.STATUS_PLANNED);
-    manufOrder.setCancelReason(null);
-    manufOrder.setCancelReasonStr(null);
-
-    return manufOrderRepo.save(manufOrder);
+        operationOrderOutsourceService);
   }
 
   @Override
