+++ axelor-production/src/main/java/com/axelor/apps/production/rest/ManufOrderRestServiceImpl.java
@@ -22,23 +22,28 @@ import com.axelor.apps.base.AxelorException;
 import com.axelor.apps.base.db.repo.TraceBackRepository;
 import com.axelor.apps.production.db.ManufOrder;
 import com.axelor.apps.production.db.repo.ManufOrderRepository;
+import com.axelor.apps.production.service.manuforder.ManufOrderPlanService;
 import com.axelor.apps.production.service.manuforder.ManufOrderWorkflowService;
 import com.google.inject.Inject;
 
 public class ManufOrderRestServiceImpl implements ManufOrderRestService {
 
   protected ManufOrderWorkflowService manufOrderWorkflowService;
+  protected ManufOrderPlanService manufOrderPlanService;
 
   @Inject
-  public ManufOrderRestServiceImpl(ManufOrderWorkflowService manufOrderWorkflowService) {
+  public ManufOrderRestServiceImpl(
+      ManufOrderWorkflowService manufOrderWorkflowService,
+      ManufOrderPlanService manufOrderPlanService) {
     this.manufOrderWorkflowService = manufOrderWorkflowService;
+    this.manufOrderPlanService = manufOrderPlanService;
   }
 
   public void updateStatusOfManufOrder(ManufOrder manufOrder, int targetStatus)
       throws AxelorException {
     if (manufOrder.getStatusSelect() == ManufOrderRepository.STATUS_DRAFT
         && targetStatus == ManufOrderRepository.STATUS_PLANNED) {
-      manufOrderWorkflowService.plan(manufOrder);
+      manufOrderPlanService.plan(manufOrder);
     } else if (manufOrder.getStatusSelect() == ManufOrderRepository.STATUS_PLANNED
         && targetStatus == ManufOrderRepository.STATUS_IN_PROGRESS) {
       manufOrderWorkflowService.start(manufOrder);
