+++ axelor-production/src/main/java/com/axelor/apps/production/service/manuforder/ManufOrderServiceImpl.java
@@ -46,6 +46,7 @@ import com.axelor.apps.production.db.ProductionConfig;
 import com.axelor.apps.production.db.ProductionOrder;
 import com.axelor.apps.production.db.repo.BillOfMaterialRepository;
 import com.axelor.apps.production.db.repo.ManufOrderRepository;
+import com.axelor.apps.production.db.repo.ProdProcessRepository;
 import com.axelor.apps.production.db.repo.ProdProductRepository;
 import com.axelor.apps.production.exceptions.ProductionExceptionMessage;
 import com.axelor.apps.production.service.BillOfMaterialService;
@@ -107,7 +108,8 @@ public class ManufOrderServiceImpl implements ManufOrderService {
 
   protected SequenceService sequenceService;
   protected OperationOrderService operationOrderService;
-  protected ManufOrderWorkflowService manufOrderWorkflowService;
+  protected ManufOrderPlanService manufOrderPlanService;
+  protected ManufOrderCreatePurchaseOrderService manufOrderCreatePurchaseOrderService;
   protected ProductVariantService productVariantService;
   protected AppBaseService appBaseService;
   protected AppProductionService appProductionService;
@@ -125,7 +127,8 @@ public class ManufOrderServiceImpl implements ManufOrderService {
   public ManufOrderServiceImpl(
       SequenceService sequenceService,
       OperationOrderService operationOrderService,
-      ManufOrderWorkflowService manufOrderWorkflowService,
+      ManufOrderPlanService manufOrderPlanService,
+      ManufOrderCreatePurchaseOrderService manufOrderCreatePurchaseOrderService,
       ProductVariantService productVariantService,
       AppBaseService appBaseService,
       AppProductionService appProductionService,
@@ -140,7 +143,8 @@ public class ManufOrderServiceImpl implements ManufOrderService {
       BillOfMaterialService billOfMaterialService) {
     this.sequenceService = sequenceService;
     this.operationOrderService = operationOrderService;
-    this.manufOrderWorkflowService = manufOrderWorkflowService;
+    this.manufOrderPlanService = manufOrderPlanService;
+    this.manufOrderCreatePurchaseOrderService = manufOrderCreatePurchaseOrderService;
     this.productVariantService = productVariantService;
     this.appBaseService = appBaseService;
     this.appProductionService = appProductionService;
@@ -238,8 +242,8 @@ public class ManufOrderServiceImpl implements ManufOrderService {
             && appProductionService.getAppProduction().getAutoPlanManufOrderFromSO()
         || manufOrderOrigin.equals(ManufOrderOriginTypeProduction.ORIGIN_TYPE_MRP)
         || manufOrderOrigin.equals(ManufOrderOriginTypeProduction.ORIGIN_TYPE_OTHER)) {
-      manufOrder = manufOrderWorkflowService.plan(manufOrder);
-      manufOrderWorkflowService.createPurchaseOrders(manufOrder);
+      manufOrder = manufOrderPlanService.plan(manufOrder);
+      manufOrderCreatePurchaseOrderService.createPurchaseOrders(manufOrder);
     }
 
     return manufOrderRepo.save(manufOrder);
@@ -570,7 +574,7 @@ public class ManufOrderServiceImpl implements ManufOrderService {
     this.createToProduceProdProductList(manufOrder);
     updateRealQty(manufOrder, manufOrder.getQty());
     LocalDateTime plannedStartDateT = manufOrder.getPlannedStartDateT();
-    manufOrderWorkflowService.updatePlannedDates(
+    manufOrderPlanService.updatePlannedDates(
         manufOrder,
         plannedStartDateT != null
             ? plannedStartDateT
@@ -1139,7 +1143,7 @@ public class ManufOrderServiceImpl implements ManufOrderService {
      */
     if (appProductionService.isApp("production")
         && appProductionService.getAppProduction().getIsManufOrderPlannedAfterMerge()) {
-      manufOrderWorkflowService.plan(mergedManufOrder);
+      manufOrderPlanService.plan(mergedManufOrder);
     } else {
       preFillOperations(mergedManufOrder);
     }
@@ -1424,7 +1428,20 @@ public class ManufOrderServiceImpl implements ManufOrderService {
   @Transactional(rollbackOn = {Exception.class})
   public void updatePlannedDates(ManufOrder manufOrder) {
 
-    manufOrder.setPlannedStartDateT(manufOrderWorkflowService.computePlannedStartDateT(manufOrder));
-    manufOrder.setPlannedEndDateT(manufOrderWorkflowService.computePlannedEndDateT(manufOrder));
+    manufOrder.setPlannedStartDateT(manufOrderPlanService.computePlannedStartDateT(manufOrder));
+    manufOrder.setPlannedEndDateT(manufOrderPlanService.computePlannedEndDateT(manufOrder));
+  }
+
+  @Override
+  public void checkApplicableManufOrder(ManufOrder manufOrder) throws AxelorException {
+    if (manufOrder.getBillOfMaterial().getStatusSelect()
+            != BillOfMaterialRepository.STATUS_APPLICABLE
+        || manufOrder.getProdProcess().getStatusSelect()
+            != ProdProcessRepository.STATUS_APPLICABLE) {
+      throw new AxelorException(
+          manufOrder,
+          TraceBackRepository.CATEGORY_INCONSISTENCY,
+          I18n.get(ProductionExceptionMessage.CHECK_BOM_AND_PROD_PROCESS));
+    }
   }
 }
